Move the expired notice into PurchaseNotice
